Properly add OIDs for `image` tags
`image` tags were lacking a style parameter and still had the image's
URL in the xlink:href, so they restyled `image` still showed the
original content and not the OID pattern. So we restyle such images to
hand-crafted rects with the proper style applied. That way, `desc`s
directly on an image now get properly restyled.