Consider only first non-tttool YAML file
When running Travis CI tests, we bring in tttool and unpack it in the
tests. Since the current tttool zip comes with a few yaml files, this
confused our logic to determine the YAML file that is to build.

So we exclude the tttool YAML files, and only pick the first of the
remaining YAML files. This way, the Travis CI tests should pass again.